rename package exceptions -> exception
