+++ rating-service/src/main/java/by/ikrotsyuk/bsuir/ratingservice/exception/GlobalExceptionHandler.java
@@ -1,9 +1,9 @@
-package by.ikrotsyuk.bsuir.ratingservice.exceptions;
+package by.ikrotsyuk.bsuir.ratingservice.exception;
 
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.dto.ExceptionDTO;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions.*;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.keys.GeneralExceptionMessageKeys;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.template.ExceptionTemplate;
+import by.ikrotsyuk.bsuir.ratingservice.exception.dto.ExceptionDTO;
+import by.ikrotsyuk.bsuir.ratingservice.exception.exceptions.*;
+import by.ikrotsyuk.bsuir.ratingservice.exception.keys.GeneralExceptionMessageKeys;
+import by.ikrotsyuk.bsuir.ratingservice.exception.template.ExceptionTemplate;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
