+++ rating-service/src/main/java/by/ikrotsyuk/bsuir/ratingservice/exception/exceptions/FeignDeserializationException.java
@@ -1,0 +1,10 @@
+package by.ikrotsyuk.bsuir.ratingservice.exception.exceptions;
+
+import by.ikrotsyuk.bsuir.ratingservice.exception.keys.GeneralExceptionMessageKeys;
+import by.ikrotsyuk.bsuir.ratingservice.exception.template.ExceptionTemplate;
+
+public class FeignDeserializationException extends ExceptionTemplate {
+    public FeignDeserializationException() {
+        super(GeneralExceptionMessageKeys.FEIGN_DESERIALIZATION_MESSAGE_KEY.getMessageKey());
+    }
+}
