+++ rating-service/src/main/java/by/ikrotsyuk/bsuir/ratingservice/exception/exceptions/RideNotBelongToPassengerException.java
@@ -1,7 +1,7 @@
-package by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions;
+package by.ikrotsyuk.bsuir.ratingservice.exception.exceptions;
 
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.keys.GeneralExceptionMessageKeys;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.template.ExceptionTemplate;
+import by.ikrotsyuk.bsuir.ratingservice.exception.keys.GeneralExceptionMessageKeys;
+import by.ikrotsyuk.bsuir.ratingservice.exception.template.ExceptionTemplate;
 
 public class RideNotBelongToPassengerException extends ExceptionTemplate {
     public RideNotBelongToPassengerException(Long rideId, Long passengerId) {
