--- rating-service/src/main/java/by/ikrotsyuk/bsuir/ratingservice/exceptions/exceptions/FeignDeserializationException.java
@@ -1,10 +1,0 @@
-package by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions;
-
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.keys.GeneralExceptionMessageKeys;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.template.ExceptionTemplate;
-
-public class FeignDeserializationException extends ExceptionTemplate {
-    public FeignDeserializationException() {
-        super(GeneralExceptionMessageKeys.FEIGN_DESERIALIZATION_MESSAGE_KEY.getMessageKey());
-    }
-}
