+++ rating-service/src/main/java/by/ikrotsyuk/bsuir/ratingservice/service/impl/RatingReviewerServiceImpl.java
@@ -6,10 +6,10 @@ import by.ikrotsyuk.bsuir.ratingservice.dto.RatingResponseDTO;
 import by.ikrotsyuk.bsuir.ratingservice.dto.feign.RideFullResponseDTO;
 import by.ikrotsyuk.bsuir.ratingservice.entity.RatingEntity;
 import by.ikrotsyuk.bsuir.ratingservice.entity.customtypes.ReviewerTypes;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions.IdIsNotValidException;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions.ReviewAlreadyExistsException;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions.ReviewNotFoundByIdException;
-import by.ikrotsyuk.bsuir.ratingservice.exceptions.exceptions.ReviewsNotFoundException;
+import by.ikrotsyuk.bsuir.ratingservice.exception.exceptions.IdIsNotValidException;
+import by.ikrotsyuk.bsuir.ratingservice.exception.exceptions.ReviewAlreadyExistsException;
+import by.ikrotsyuk.bsuir.ratingservice.exception.exceptions.ReviewNotFoundByIdException;
+import by.ikrotsyuk.bsuir.ratingservice.exception.exceptions.ReviewsNotFoundException;
 import by.ikrotsyuk.bsuir.ratingservice.feign.RideClient;
 import by.ikrotsyuk.bsuir.ratingservice.kafka.producer.RatingProducer;
 import by.ikrotsyuk.bsuir.ratingservice.mapper.RatingMapper;
